refactor: Rename `lib/setApiGatewayAccessLogFormat.js`
