+++ lib/plugin.js
@@ -19,7 +19,7 @@ const { generate } = require('./generate-event');
 const { configureDeployProfile } = require('./deploy-profile');
 const { test } = require('./test');
 const { getDashboardUrl, dashboardHandler } = require('./dashboard');
-const setApiGatewayAccessLogFormat = require('./setApiGatewayAccessLogFormat');
+const setApiGatewayAccessLogFormat = require('./set-api-gateway-access-log-format');
 const paramCommand = require('./param-command');
 const outputCommand = require('./output-command');
 const isAuthenticated = require('./is-authenticated');
