+++ lib/plugin.test.js
@@ -52,7 +52,7 @@ describe('plugin', () => {
       './deploy-profile': { configureDeployProfile },
       './inject-logs-iam-role': injectLogsIamRole,
       './is-authenticated': () => true,
-      './setApiGatewayAccessLogFormat': setApiGatewayAccessLogFormat,
+      './set-api-gateway-access-log-format': setApiGatewayAccessLogFormat,
       '@serverless/config/utils': {
         getLoggedInUser: () => ({
           accessKeys: {
