+++ lib/set-api-gateway-access-log-format.js
