+++ lib/set-api-gateway-access-log-format.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { expect } = require('chai');
-const setApiGatewayAccessLogFormat = require('./setApiGatewayAccessLogFormat');
+const setApiGatewayAccessLogFormat = require('./set-api-gateway-access-log-format');
 const { API_GATEWAY_LOG_FORMAT, API_GATEWAY_V2_LOG_FORMAT } = require('./utils');
 
 describe('setApiGatewayAccessLogFormat', () => {
